src/requirements.txt
async-generator==1.10
bcrypt==3.2.2
easydict==1.9
fastapi==0.78.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gunicorn==20.1.0
Imageio==2.15.0
ipython==8.0.1
Jinja2==3.1.2
joblib==1.1.0
jupyter-client==7.1.2
jupyter-core==4.9.1
oauthlib==3.1.1
passlib==1.7.4
psycopg2==2.9.3
psycopg2-binary==2.9.3
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
qrcode==7.3.1
requests==2.27.1
requests-oauthlib==1.3.0
SQLAlchemy==1.4.39
starlette==0.19.1
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.18.2
virtualenv==20.15.1
